requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
tldextract==5.1.2
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-groq==0.2.1
jinja2==3.1.6
